+++ packages/solid/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Rename TypeScript types and export more of them
+- Fix typo that causes bugs when using field arrays
 
 ## v0.5.0 (November 06, 2022)